support user defined read/create/update/delete timeouts for all resources in the provider (#408)
* chore: adding local provider setup for acceptance testing

* test

* remove test code

* deps fix

* adding terraform setup

* running deps logic

* fixing script issues

* adding config file export var

* adding pr trigger for testing

* testing acc tests

* removing debug msgs

* testing pipeline

* runnign acceptance tests

* testing pipeline

* adding oidc login for azuread

* cleaning up tests to use 3rd party providers correctly

* moving tests to services subdirectories and renaming packages for all services

* running unit tests

* renaming files

* fixing artifact names

* validating tests

* typo

* testing acceptance test in application service

* validation azure permissions

* //TODO: turning off until we fix the testing tenant

* testing TestAccBillingPolicyResource_Validate_Create

* testing TestAccBillingPolicyResource_Validate_Create

* removing azurecaf from test

* adding subscription id env var  to the pipeline

* billing policy name too long fix

* fixing policy name test

* testing TestAccBillingPolicyResourceEnvironment_Validate_Create

* testing TestAccSolutionResource_Validate_Create_With_Settings_File

* testign TestAccEnvironmentApplicationPackageInstallResource_Validate_Install

* running all acceptance tests

* fixing TestAccEnvironmentPowerAppsDataSource_Basic test

* fixing TestAccEnvironmentPowerAppsDataSource_Basic

* testing TestAccSecurityDataSource_Validate_Read

* typo fix

* adding debug trace

* testing sleep in TestAccSecurityDataSource_Validate_Read

* adding debug msg

* testing TestAccDataRecordDatasource_Validate_Expand_Query

* testing TestAccTestEnvironmentSettingsResource_Validate_Read

* testing TestAccTestEnvironmentSettingsResource_Validate_Read

* testing TestAccDataRecordDatasource_Validate_Expand_Query

* testing TestAccDataRecordDatasource_Validate_Expand_Query

* testing TestAccTestRest_Validate_Create

* testing TestAccSolutionsDataSource_Validate_Read

* testing acceptance tests

* fixing 401 issues

* set parallel to 1

* moving all acceptance test to run in unitedstates region

* running acceptance tests

* fixing TestAccTestRest_Validate_Create and testing p=2

* fixing TestAccDataRecordDatasource_Validate_UserQuery

* update docs

* adding wait for TestAccDataRecordResource_Validate_Create

* testing TestAccEnvironmentPowerAppsDataSource_Basic

* testing with p=3

* adding 60sec wait to all test that need to change newly created environment

* testing TestAccBillingPolicyResourceEnvironment_Validate_Update

* testing licensing package

* TestAccBillingPoliciesEnvironmentsDataSource_Validate_Read fix

* TestAccBillingPoliciesEnvironmentsDataSource_Validate_Read fix

* fix TestAccBillingPolicy_Validate_Update

* running all test with p=3

* running all test with p=3

* p=2

* disble TestAccDlpPolicyDataSource_Validate_Read

* fix

* adding retry on env creation

* removing wait resource

* testing p=6

* removing comments

* adding retry for all requests

* fixing tests and running p=12

* testingTestAccTestRest_Validate_Create

* fixing unit tests

* test TestAccTestRest_Validate_Create

* running all acceptance tests

* testing TestAccUserResource_Validate_Create

* testing TestAccConnectionsShareDataSource_Validate_Read

* testing TestAccConnectionsShareResource_Validate_Create

* running all tests

* fixing dev container

* adding delete timeout to billing policy and refactoring timeout as const

* refactoring constants package name

* removing explicit package aliases

* adding update/create/delete timeout to rest of the resources

* update docs

* running PR only on go.mod changes

* unit test p=16

* chore: Add timeouts to resource models and update documentation

* adding read timeout

* testing TestAccTestRest_Validate_Create

* running all tests

* changing run logic

* testing TestAccTestRest_Validate_Create

* chore: Refactor logging in SendOperation method

* running all acc tests

* fixing code after merge

* docs regen

* turning off acceptance tests for PR

* turning off acc tests

* Update .github/workflows/run_tests.yml


* review fixes

* Update run_tests.yml

* merge mawasile/375-new-tenant-for-acceptance-tests

* solution fix

* removing unused param

* consolidation retryAfter time logic

* consolidating timeout logic

* adding more statues that would raise exception on HTTP response

---------

+++ internal/powerplatform/services/connection/resource_connection_share.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/hashicorp/terraform-plugin-framework-timeouts/resource/timeouts"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -16,7 +17,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	api "github.com/microsoft/terraform-provider-power-platform/internal/powerplatform/api"
+	"github.com/microsoft/terraform-provider-power-platform/internal/powerplatform/api"
+	"github.com/microsoft/terraform-provider-power-platform/internal/powerplatform/constants"
 )
 
 var _ resource.Resource = &ConnectionShareResource{}
@@ -36,6 +38,7 @@ type ConnectionShareResource struct {
 }
 
 type ConnectionShareResourceModel struct {
+	Timeouts      timeouts.Value                        `tfsdk:"timeouts"`
 	Id            types.String                          `tfsdk:"id"`
 	EnvironmentId types.String                          `tfsdk:"environment_id"`
 	ConnectorName types.String                          `tfsdk:"connector_name"`
@@ -57,6 +60,12 @@ func (r *ConnectionShareResource) Schema(ctx context.Context, req resource.Schem
 	resp.Schema = schema.Schema{
 		MarkdownDescription: "",
 		Attributes: map[string]schema.Attribute{
+			"timeouts": timeouts.Attributes(ctx, timeouts.Opts{
+				Create: true,
+				Update: true,
+				Delete: true,
+				Read:   true,
+			}),
 			"id": schema.StringAttribute{
 				MarkdownDescription: "Unique identifier of the connection share",
 				Computed:            true,
@@ -146,6 +155,15 @@ func (r *ConnectionShareResource) Create(ctx context.Context, req resource.Creat
 		return
 	}
 
+	timeout, diags := plan.Timeouts.Create(ctx, constants.DEFAULT_RESOURCE_OPERATION_TIMEOUT_IN_MINUTES)
+	if diags != nil {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
 	err := r.ConnectionsClient.ShareConnection(ctx, plan.EnvironmentId.ValueString(), plan.ConnectorName.ValueString(), plan.ConnectionId.ValueString(), plan.RoleName.ValueString(), plan.Principal.EntraObjectId.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error sharing connection", err.Error())
@@ -178,6 +196,15 @@ func (r *ConnectionShareResource) Read(ctx context.Context, req resource.ReadReq
 	if resp.Diagnostics.HasError() {
 		return
 	}
+
+	timeout, diags := state.Timeouts.Read(ctx, constants.DEFAULT_RESOURCE_OPERATION_TIMEOUT_IN_MINUTES)
+	if diags != nil {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
 
 	share, err := r.ConnectionsClient.GetConnectionShare(ctx, state.EnvironmentId.ValueString(), state.ConnectorName.ValueString(), state.ConnectionId.ValueString(), state.Principal.EntraObjectId.ValueString())
 	if err != nil {
@@ -225,6 +252,16 @@ func (r *ConnectionShareResource) Update(ctx context.Context, req resource.Updat
 		},
 		Delete: []ShareConnectionRequestDeleteDto{},
 	}
+
+	timeout, diags := plan.Timeouts.Update(ctx, constants.DEFAULT_RESOURCE_OPERATION_TIMEOUT_IN_MINUTES)
+	if diags != nil {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
 	err := r.ConnectionsClient.UpdateConnectionShare(ctx, plan.EnvironmentId.ValueString(), plan.ConnectorName.ValueString(), plan.ConnectionId.ValueString(), share)
 	if err != nil {
 		resp.Diagnostics.AddError("Error updating connection share", err.Error())
@@ -259,6 +296,15 @@ func (r *ConnectionShareResource) Delete(ctx context.Context, req resource.Delet
 		return
 	}
 
+	timeout, diags := state.Timeouts.Delete(ctx, constants.DEFAULT_RESOURCE_OPERATION_TIMEOUT_IN_MINUTES)
+	if diags != nil {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
 	err := r.ConnectionsClient.DeleteConnectionShare(ctx, state.EnvironmentId.ValueString(), state.ConnectorName.ValueString(), state.ConnectionId.ValueString(), state.Id.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error deleting connection share", err.Error())
@@ -272,6 +318,7 @@ func (r *ConnectionShareResource) ImportState(ctx context.Context, req resource.
 
 func ConvertFromConnectionResourceSharesDto(oldPlan *ConnectionShareResourceModel, connection *ShareConnectionResponseDto) ConnectionShareResourceModel {
 	share := ConnectionShareResourceModel{
+		Timeouts:      oldPlan.Timeouts,
 		EnvironmentId: oldPlan.EnvironmentId,
 		ConnectorName: oldPlan.ConnectorName,
 		ConnectionId:  oldPlan.ConnectionId,
